Configure inside django app
